requirements.txt
absl-py==1.3.0
accelerate
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.8
appdirs==1.4.4
astor==0.8.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
backcall
beautifulsoup4
bitarray==2.7.3
bleach
blis==0.7.8
blobfile==2.0.0
brotlipy
cachetools==5.2.0
catalogue
certifi==2023.5.7
cffi
charset-normalizer
chumpy==0.70
click==8.1.3
clip
cloudpickle==2.2.1
cmake==3.26.1
colorama
commonmark==0.9.1
confection==0.0.2
ConfigArgParse==1.5.3
contextlib2==21.6.0
cryptography
cycler
cymem
Cython==0.29.34
dataclasses
decorator
defusedxml
docker-pycreds==0.4.0
easydict==1.10
einops==0.6.0
entrypoints
essentia==2.1b6.dev1034
fairseq==0.12.2
fastjsonschema
filelock
fire==0.1.3
flit_core
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
future
fvcore==0.1.5.post20221221
gdown
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
h5py
hydra-core==1.0.7
idna
importlib-metadata==5.0.0
importlib-resources==5.12.0
iopath==0.1.10
ipykernel
ipython
ipython-genutils==0.2.0
ipywidgets
jedi
Jinja2
joblib
json-tricks==3.16.1
jsonschema==2.6.0
jupyter
langcodes
librosa==0.9.2
lightning-utilities==0.8.0
lit==16.0.0
llvmlite==0.39.1
lmdb==1.3.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe
matplotlib==3.1.3
matplotlib-inline
mistune
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
multidict==6.0.4
murmurhash==1.0.8
nbclient
nbconvert
nbformat
nest-asyncio
nni==2.10
notebook
numba==0.56.4
numpy
oauthlib==3.2.2
omegaconf==2.0.6
onnx
opencv-python==4.6.0.66
packaging
pandas
pandocfilters
parso
pathtools==0.1.2
pathy
pexpect
pickleshare
Pillow==9.2.0
pooch==1.6.0
portalocker==2.7.0
praat-parselmouth==0.4.3
prefetch-generator==1.0.3
preshed==3.0.7
prettytable==3.6.0
prometheus-client
prompt-toolkit
psutil==5.9.4
ptyprocess
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser
pycryptodomex==3.15.0
pydantic
pydub==0.25.1
Pygments
pyloudnorm==0.1.0
pyOpenSSL
pyparsing
PySocks
PythonWebHDFS==0.2.3
pytorch-lightning
pytube==12.1.2
pytz==2022.7
PyYAML
pyzmq==19.0.2
regex==2022.9.13
requests
requests-oauthlib==1.3.1
resampy==0.4.2
responses==0.23.1
rich==12.5.1
rsa==4.9
sacrebleu==2.3.1
schema==0.7.5
scikit-learn
scipy
seaborn==0.12.2
sentry-sdk==1.22.2
setproctitle==1.3.2
simplejson==3.18.4
six
smmap==5.0.0
smplx==0.1.28
SoundFile==0.10.3.post1
sox==1.4.1
spacy
srsly==2.4.4
tabulate==0.9.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
TextGrid==1.5
thinc==8.0.17
torch==1.12.1
torchaudio
torchmetrics
torchvision
tqdm
transforms3d==0.4.1
triton==2.0.0.post1
types-PyYAML==6.0.12.9
Unidecode==1.1.1
wandb==0.15.2
webencodings==0.5.1
websockets==11.0.2
Werkzeug==2.2.2
wget==3.2
yacs==0.1.8
yarl==1.8.2
zipp
